Company Details module added

+++ medical_store/models.py
@@ -24,7 +24,7 @@ class MedicineInventory(models.Model):
 
 class CompanyDetails(models.Model):
     company_name = models.CharField(max_length=200)
-    company_contact = models.PositiveIntegerField(validators=[MaxValueValidator(9999999999),MinValueValidator(1000000000)])
+    company_contact = models.BigIntegerField(validators=[MaxValueValidator(9999999999),MinValueValidator(1000000000)])
     company_address=models.CharField(max_length=200)
     company_email = models.EmailField()
     gst_number = models.CharField(max_length=15)
